Split numautils into header and source to avoid including windows.h into all the code
